Respostas para cada requisição criadas

+++ src/main/kotlin/com/acme/tour/controller/PromocaoController.kt
@@ -1,10 +1,13 @@
 package com.acme.tour.controller
 
 import com.acme.tour.model.Promocao
+import com.acme.tour.model.RespostaJSON
 import com.acme.tour.service.PromocaoService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import java.util.concurrent.ConcurrentHashMap
+import java.util.*
 
 @RestController
 @RequestMapping(value = ["/promocoes"])
@@ -14,26 +17,50 @@ class PromocaoController {
     lateinit var promocaoService: PromocaoService
 
     @GetMapping("/{id}")
-    fun getGetById(@PathVariable id: Long) = this.promocaoService.getById(id)
+    fun getGetById(@PathVariable id: Long): ResponseEntity<Promocao?> {
+        val promocao = this.promocaoService.getById(id)
+        val status = if (promocao == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+        return ResponseEntity(promocao, status)
+    }
 
     @PostMapping()
-    fun create(@RequestBody promocao: Promocao) {
+    fun create(@RequestBody promocao: Promocao): ResponseEntity<RespostaJSON> {
         this.promocaoService.create(promocao)
+        val respostaJSON = RespostaJSON("Ok", Date())
+        return ResponseEntity(respostaJSON, HttpStatus.CREATED)
     }
 
     @DeleteMapping("/{id}")
-    fun delete(@PathVariable id: Long) {
-        this.promocaoService.delete(id)
+    fun delete(@PathVariable id: Long): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (this.promocaoService.getById(id) != null) {
+            status = HttpStatus.ACCEPTED
+            this.promocaoService.delete(id)
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
     @PutMapping("/{id}")
-    fun update(@PathVariable id: Long, @RequestBody promocao: Promocao) {
-        this.promocaoService.update(id, promocao)
+    fun update(@PathVariable id: Long, @RequestBody promocao: Promocao): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (this.promocaoService.getById(id) != null) {
+            this.promocaoService.update(id, promocao)
+            status = HttpStatus.ACCEPTED
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
     //@RequestMapping(value = ["/promocoes"], method = arrayOf(RequestMethod.GET))
     @GetMapping()
-    fun getAll(@RequestParam(required = false, defaultValue = "") localFilter: String) =
-        this.promocaoService.searchByLocal(localFilter)
+    fun getAll(@RequestParam(required = false, defaultValue = "") localFilter: String): ResponseEntity<List<Promocao>> {
+        var status = HttpStatus.OK
+        val listaPromocoes = this.promocaoService.searchByLocal(localFilter)
+        if (listaPromocoes.isEmpty()) {
+            status = HttpStatus.NOT_FOUND
+        }
 
+        return ResponseEntity(listaPromocoes, status)
+    }
 }
